merge, fix toast click, refactor to queueFormatter
