+++ client/src/components/NewExperiment.vue
@@ -9,7 +9,7 @@ import { store, pollForResult } from '../store.js'
 import { API_URL } from '../api'
 import { emptyOption } from '../helpers/optionsFormatter'
 import { emptyFormGroup, validateEmail } from '../helpers/optionsFormatter'
-import { progress } from '../helpers/resultFormatter'
+import { progress } from '../helpers/queueFormatter'
 
 const horizontalLayout = ref(false)
 const oldMetadata = ref(false)
@@ -265,7 +265,6 @@ function reformat(property) {
                   <b-col md="auto">
                     <!--User can select the amount of recommendations per user -->
                     <b-form-group
-                      
                       label="Select number of recommendations per user: *"
                     >
                       <b-form-input
